Updates for database export

+++ code/web/cron/backupAspen.php
@@ -22,16 +22,20 @@ if (!file_exists("/data/aspen-discovery/$serverName/sql_backup")) {
 
 //Remove any backups older than 3 days
 $backupDir = "/data/aspen-discovery/$serverName/sql_backup";
-exec_advanced("find $backupDir/ -mindepth 1 -maxdepth 1 -name *.sql -type f -mtime +3 -delete", $debug);
-exec_advanced("find $backupDir/ -mindepth 1 -maxdepth 1 -name *.sql.gz -type f -mtime +3 -delete", $debug);
-exec_advanced("find $backupDir/ -mindepth 1 -maxdepth 1 -name *.tar -type f -mtime +3 -delete", $debug);
-exec_advanced("find $backupDir/ -mindepth 1 -maxdepth 1 -name *.tar.gz -type f -mtime +3 -delete", $debug);
+if ($configArray['System']['operatingSystem'] != 'windows') {
+	exec_advanced("find $backupDir/ -mindepth 1 -maxdepth 1 -name *.sql -type f -mtime +3 -delete", $debug);
+	exec_advanced("find $backupDir/ -mindepth 1 -maxdepth 1 -name *.sql.gz -type f -mtime +3 -delete", $debug);
+	exec_advanced("find $backupDir/ -mindepth 1 -maxdepth 1 -name *.tar -type f -mtime +3 -delete", $debug);
+	exec_advanced("find $backupDir/ -mindepth 1 -maxdepth 1 -name *.tar.gz -type f -mtime +3 -delete", $debug);
+}
 
 //Create the tar file
 $curDateTime = date('ymdHis');
 $backupFile = "$backupDir/aspen.$serverName.$curDateTime.tar";
 //exec("tar -cf $backupFile");
-exec_advanced("cd $backupDir", $debug);
+if ($configArray['System']['operatingSystem'] != 'windows') {
+	exec_advanced("cd $backupDir", $debug);
+}
 
 //Create the export files
 $listTablesStmt = $aspen_db->query("SHOW TABLES");
@@ -39,31 +43,80 @@ $allTables = $listTablesStmt->fetchAll(PDO::FETCH_COLUMN);
 foreach ($allTables as $table) {
 	$exportData = true;
 	//Ignore
-	if ($table == 'session' || $table == 'cached_values') {
+	if ($table == 'session' || $table == 'cached_values' || $table == 'external_request_log') {
 		$exportData = false;
 	}
 
 	$exportFile = "$serverName.$curDateTime.$table.sql";
 	$fullExportFilePath = "$backupDir/$exportFile";
-	if ($exportData) {
-		$dumpCommand = "mariadb-dump --quick -u$dbUser -p$dbPassword -h$dbHost -P$dbPort $dbName $table > $fullExportFilePath";
-	}else{
-		$dumpCommand = "mariadb-dump --quick -u$dbUser -p$dbPassword -h$dbHost -P$dbPort --no-data $dbName $table > $fullExportFilePath";
+
+	$createTableStmt = $aspen_db->query("SHOW CREATE TABLE " . $table);
+	$createTablesRS = $createTableStmt->fetchAll(PDO::FETCH_ASSOC);
+	$fhnd = fopen($fullExportFilePath, 'a+');
+	fwrite($fhnd, "DROP TABLE IF EXISTS $table;\n");
+	foreach ($createTablesRS as $createTableSql) {
+		$createTableValue = $createTableSql['Create Table'];
+		//Remove the auto increment id
+		$createTableValue = preg_replace('/AUTO_INCREMENT=\d+/', '', $createTableValue);
+		fwrite($fhnd, $createTableValue . ";\n");
 	}
-	exec_advanced($dumpCommand, $debug);
+	$createTableStmt->closeCursor();
+
+	if ($exportData) {
+		$exportDataStmt = $aspen_db->query("SELECT * FROM " . $table);
+		$isFirstRow = true;
+		$hasData = false;
+		while ($row = $exportDataStmt->fetch(PDO::FETCH_ASSOC)) {
+			$hasData = true;
+			if ($isFirstRow) {
+				$columns = implode(',', array_keys($row));
+				$insertStatement = "INSERT INTO $table ($columns) VALUES ";
+				fwrite($fhnd, $insertStatement);
+			}
+			$values = [];
+			foreach ($row as $value) {
+				if (is_numeric($value)) {
+					$values[] = $value;
+				}else{
+					$values[] = "'" . str_replace("'", "/'", $value) . "'";
+				}
+			}
+			if (!$isFirstRow) {
+				fwrite($fhnd, ", ");
+			}
+			fwrite($fhnd, "(" . implode(',', $values) . ")");
+
+			$isFirstRow = false;
+		}
+		if ($hasData) {
+			fwrite($fhnd, ";\n");
+		}
+		$exportDataStmt->closeCursor();
+	}
+	fclose($fhnd);
+//	if ($exportData) {
+//		$dumpCommand = "mariadb-dump --quick -u$dbUser -p$dbPassword -h$dbHost -P$dbPort $dbName $table > $fullExportFilePath";
+//	}else{
+//		$dumpCommand = "mariadb-dump --quick -u$dbUser -p$dbPassword -h$dbHost -P$dbPort --no-data $dbName $table > $fullExportFilePath";
+//	}
+//	exec_advanced($dumpCommand, $debug);
 
 	//remove the exported file
 	if (file_exists($fullExportFilePath)) {
 		//Add the file to the archive
-		exec_advanced("cd $backupDir; tar -rf $backupFile $exportFile", $debug);
+		if ($configArray['System']['operatingSystem'] != 'windows') {
+			exec_advanced("cd $backupDir; tar -rf $backupFile $exportFile", $debug);
 
-		unlink($fullExportFilePath);
+			unlink($fullExportFilePath);
+		}
 	}
 
 }
 
 //zip up the archive
-exec_advanced("gzip $backupFile", $debug);
+if ($configArray['System']['operatingSystem'] != 'windows') {
+	exec_advanced("gzip $backupFile", $debug);
+}
 
 //Optionally move the file to the Google backup bucket
 // Load the system settings
